Add files via upload
virtual-assistant-Desktop-
The project aims to develop a personal-assistant for windowsbased systemsJarvis draws its inspiration from virtual assistants like Cortana for Windows, and Siri for iOS. It has been designed to provide a user-friendly interface for carrying out a variety of tasks by employing certain well-defined commands. Users can interact with the assistant either through voice commands or using keyboard input. As a personal assistant, Jarvis assists the end-user with day-to-day activities like general human conversation, searching queries in google, bing or yahoo, searching for videos, retrieving images, live weather conditions, word meanings, searching for medicine details, reminding the user about the scheduled events and tasks. The user statements/commands are analysed with the help of machine learning to give an optimal solution. Keywords:- Personal Assistant, , Automation, Machine Learning.

Queries from the web: Making queries is an essential part of one’s life, and nothing changes even for a developer working on Linux. We have addressed the essential part of a netizen’s life by enabling our voice assistant to search the web. Here we have used Node JS and Selenium framework for extracting the result from the web as well as displaying it to the user. Jarvis supports a plethora of search engines like Google, Bing and Yahoo and displays the result by scraping the searched queries. In order to make queries from different search engines, the given format should be adopted:
Accessing youtube videos : Videos have remained as a main source of entertainment, one of the most prioritized tasks of virtual assistants. They are equally important for entertainment as well as educational purposes as most teaching and research activities in present times are done through Youtube. This helps in making the learning process more practical and out of the four walls of the classroom. Jarvis implements the feature through a subprocess module which is handled by the main Golang service. This service initiates the subprocess for Node JS which serves the Selenium WebDriver, and scraps the searched YouTube query.
Get weather for a location: Getting live weather conditions about a place remains an important task of virtual assistants. It helps the user charter the course of their action. Jarvis addresses this issue with the help of Python.
Set Reminders : One of the main features of a voice assistant is to set a reminder for the user accordingly. Jarvis is no different when it comes to this. The user can set reminders to be notified about a task at a particular time. This will help users, especially developers to schedule their time and resources easily. All the user have to do is to input Set reminder to the assistant. A form will be displayed. Fill the form with the required details and click on set reminder button.
Sending Emails : Integrating mailing features to Jarvis eases the job of mailing, which otherwise would have to be done by opening the concerned email address. With Jarvis, you do not need to go for another tab to do one of the major task of your day to day affairs. The user can send emails to the desired receiver. He should input Send mail, after which a form will be displayed. Fill the form with the required details and click on the send mail button. And it can also perform many functions like Corona update, Bank Information,Air Quality of any city,Sunrise & Sunset Timing of any city and Many more.